# apd: optimize Quo, remove per-digit long division 